pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawchat-gateway"
version = "0.1.0"
description = "ClawChat v2 gateway adapter for hermes-agent"
requires-python = ">=3.11"
dependencies = ["websockets>=12,<16", "PyYAML>=6,<7"]

[project.scripts]
clawchat-gateway-install = "clawchat_gateway.install:main"

[project.optional-dependencies]
test = ["pytest>=8", "pytest-asyncio>=0.23", "pytest-cov>=4"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["."]
